Rename jsi/hermes.h to jsi/hermes-interfaces.h (#1844)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/1844

Pull Request resolved: https://github.com/facebook/hermes/pull/1843

X-link: https://github.com/facebook/react-native/pull/54569

Changelog: [Internal]

Two "hermes.h" under different prefixes (jsi/hermes.h and
hermes/hermes.h) are a bit confusing. Let's rename the JSI
header to "hermes-interfaces.h".

Reviewed By: tmikov

Differential Revision: D87200512

fbshipit-source-id: e61251f791788c8d6787ee29762c4aa76d99b68a
