+++ API/hermes/hermes.h
@@ -18,7 +18,7 @@
 #include <hermes/Public/HermesExport.h>
 #include <hermes/Public/RuntimeConfig.h>
 #include <hermes/Public/SamplingProfiler.h>
-#include <jsi/hermes.h>
+#include <jsi/hermes-interfaces.h>
 #include <jsi/jsi.h>
 #include <unordered_map>
